requirements.txt
agora-token-builder==1.0.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
amqp==5.1.1
anyio==4.0.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
atomicwrites==1.4.1
attrs==23.1.0
auto-py-to-exe==2.38.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
bottle==0.12.25
bottle-websocket==0.2.9
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.0
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.3
cssselect2==0.7.0
customtkinter==5.2.0
cycler==0.11.0
dadata==21.10.1
daphne==4.0.0
darkdetect==0.8.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==2.2.16
django-allauth==0.53.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-channels==0.7.0
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-faicon==0.1.0
django-ipware==5.0.1
django-jazzmin==2.6.0
django-jet==1.0.8
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-nested-admin==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-timezone-field==6.1.0
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dnspython==2.4.2
Eel==0.16.0
environs==9.5.0
eventlet==0.33.3
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.1.1
fastjsonschema==2.18.0
filelock==3.12.3
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
gevent==23.9.0.post1
gevent-websocket==0.10.1
googletrans==4.0.0rc1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.10.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
kombu==5.3.2
lxml==4.9.3
magic-filter==1.0.9
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
msgpack==1.0.5
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
oscrypto==1.3.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==0.13.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyJWT==2.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyPDF3==1.0.6
pypng==0.20220715.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python-json-logger==2.0.7
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
redis==4.6.0
referencing==0.30.2
reportlab==4.0.4
requests==2.26.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.10.2
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sslcommerz-lib==1.0
stack-data==0.6.2
svglib==1.5.1
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transliterate==1.10.2
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.18.0
uri-template==1.3.0
uritools==4.0.2
urllib3==1.26.8
user-agents==2.2.0
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
whichcraft==0.6.1
whitenoise==6.2.0
widgetsnbextension==3.5.2
xhtml2pdf==0.2.8
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
